Moved linkValidation method to another directory
