--- errors/validation-err.js
@@ -1,8 +1,0 @@
-class ValidationError extends Error {
-  constructor(message) {
-    super(message);
-    this.statusCode = 400;
-  }
-}
-
-module.exports = ValidationError;
