+++ validators/linkValidator.js
